Moved Debug page to routes
